Android Dialer
The App user types a number or name in the interface. Now the app will
look for a phone number embedded in the text typed by the user.
If a phone number conforming to the a phone number format is found, the
app opens a dialer window with that phone number showing in the
corresponding window. The user may or may not decide to dial the number